fix: handle some webtiles race conditions
Most of this was found while debugging cao behavior when out of disk
space. A number of extremely strange race conditions could happen when
game startup fails in unusual ways, as well as in a subsequent attempt
to shutdown the server. This commit tries to handle these a bit better.
Unfortunately, the best solution I found for the big one (a race
condition after fork, but before execvpe overwrites python) is simply to
delay so that execvpe takes hold; fork + asyncio is not well behaved.
See:
    https://github.com/python/cpython/issues/66197
As far as I can tell, the promised solutions of safely closing the event
loop in a child never materialized.