Rework Okawaru's Duel ability somewhat
Duel was an ability with great flavor that, in practice, was usually best
used in ways that went against that flavor - by duelling relatively
harmless enemies so that the player could stall to heal up. Using it
against genuinely scary enemies often meant trapping the player in a fight
against them with no escape options and limited reward for doing so. This
aims to improve that situation by tweaking a few things about how Duel
works.

When you start a duel, the player will now instantly heal for half of their
missing health. But when the duel ends, your hp/mp will be reset to
whatever they were when the duel *started*. This means simultaneously that
there is no benefit to stalling to heal (you'll just lose that health when
the duel ends), but also that you sometimes stand to benefit greatly from
taking a risk (getting a large heal when at low hp, or undoing any battle
damage from one scary monster before you fight others)

Duels also now end instantly once the main target is dead. All items on the
floor are automatically gathered up and deposited at the player's feet,
where they exit. (In the rare case of the player standing over deep water
or lava, an Okawaru altar will be made at their feet to ensure the items
are not lost.)

Finally, the piety of cost of duel is also significantly reduced, such that
it is equivalent to Finesse.