Don't return from the Abyss into inaccessible areas
If a character with temporary flight was banished over lava or deep water,
upon departing the Abyss, they would be placed on the nearest safe tile
they could occupy - but this displacement did not respect no_tele_into and
could sometimes result in them landing in places the player was never
supposed to be able to enter (and possibly trapping them there).

This does mean we will sometimes push the player further than necessary (if
they got banished in this way from a ghost vault somehow), but that seems
more harmless than the behavior it's replacing.

(This scenario also applies to Duel, by the way)