Don't shift players arriving on stairs with no_tele_into (grumposus)
I'd thought this logic was a fallback only used if the position the player
was arriving on was unsafe, but in fact it was call every time regardless.
This meant we would shift the player merely because they were in a
no_tele_into zone, even if where they were standing was perfectly safe.

Now we should probably only care about no_tele_into if we have to move the
player at all.