Add x-v row showing primary melee damage/flavour for ranged monsters
Except for Nessos, monsters never apply an attack flavour to their
ranged attacks. Previously, x-v would pretend that monsters' ranged
weapon attacks additionally applied their primary melee attack's
flavour - this most notably applied to salamanders. Fix this.

So that the player can always see the attack flavour of a monster's
primary melee attack, add an additional row to the attacks table for
monsters that wield a ranged weapon to show the base melee damage
and flavour of the monster's primary melee attack.

Ideally, we would show the damage of the monster's alternate weapon
(if they have one) if and only if the player has already seen them
wield it. However, this information is currently not tracked by the
game, and would require a disproportionate amount of work to implement
for what is a very minor improvement.

[Note: while I believe a monster cannot dual-wield or dual-fire
ranged weapons at present, the code does not assume this, in case it
becomes possible in the future.]