Let Harpoon Shot pull monsters
This seems to have never functioned, even though the code for it clearly
assumed things other than the player *could* be pulled (eg: there was code
to check if the player could see the actor being pulled). There doesn't
seem to be any particular reason to prevent monsters from being pulled by
this (and in fact, warrior apostles can come with Harpoon Shot because I
assumed it could do something!).