Refactor collision damage and adjust several things about it.
Previously, all knockbacks/collisions which could inflict damage did so via
a fixed formula contained in actor::collide (though the power supplied to
this formula varied greatly based on source). This made it slightly
inflexible for things which wanted to do more specific things with how they
applied collision damage (ie: Gavotte or Piledriver caring about the
distance a monster has travelled before the collision). actor:collide now
instead takes a pre-AC damage number that it will apply to the colliding
actors, if applicable. This is in anticipation of a future adjustment to
Piledriver's damage, but is more flexible in general.

This commit also makes several other changes in the process:
 -Primal Wave and Mud Breath no longer inflict collision damage. This
  happened fairly rarely for Primal Wave, due to low knockback distance and
  being most common in very open areas, but could inflict up to 2d25
  additional damage when it happened (which wasn't indicated by xv!). I
  nudged the spell damage up slightly in exchange for no longer causing
  nasty surprises if the beam angle lines up just so.
  (Mud Breath, due to way in which collison damage was calculated, only did
  2d3 by XL 27...)
 -Force Lance knockback distance and chance no longer scale with
  spellpower (all current casters are pretty similar anyway)
 -Wind Blast and Force Lance collision damage should be indicated by xv
 -Brom's Barrelling Boulder inflicts no collision damage when it cannot
  displace an actor (it was only ever 2d2 anyway)
 -Force Lance (the weapon) has a higher minimum collision damage it can
  inflict (now that such things are easy to add) and overall does more
  collision damage in general. (It's one of the fun points about the weapon
  that nudges you to use it differently, and I've never heard anyone say it
  was especially strong)

Unless stated otherwise, damage from other effects should not have changed.