Change Piledriver damage formula
Piledriver asks some unique positioning questions of the player and can do
very solid damage for a low-level spell when the player lines up a max
distance shot with it. However, because it was tied to the default
collision damage formula, distance could only scale the dice *size*,
meaning that even a high power Piledriver would often wiff. This is not
unusual for spells in Crawl, but for Piledriver it was especially bad.
Lining up a good shot often takes multiple turns of setup and cannot simply
be repeated immediately (or perhaps at all), *and* often asked the player
to lunge into a more dangerous position in the process. Especially as a
starting spell of a weaker background, the payoff needs to feel more
reliable for people to invest the necessary skill points (and also do what
the spell asks of them).

To that end, instead of 2d(X-Y) damage, Piledriver now does (3-9)dX damage.
(Average damage should remain similar, and the ratio between min and max
distance shots also remains similar (4 tiles was ~3x as strong as 1 tile).
High rolls may be a little less common, but making it far less likely not
to get *a* relevant amount of damage may hopefully make it feel better to
use.