Alert monsters hit by enchantments *after* the enchantment effect
This is aimed primarily to avoid the relatively-common situation of casting
Ensorcelled Hibernation on a wandering monster that is unaware of you, and
this causing it to shout and wake up other monsters immediately, even when
the spell successfully puts it to sleep. Now (like non-enchantment beams),
the monster is alerted only *after* the effect happens - and being asleep
will stop them from shouting.

This technically buffs a similar situation with paralysis wands and maybe
a few others, but that seems fine.

I am slightly concerned this might have unforseen consequences (since it's
not obvious if there was a *reason* behavior events were fired early in
this function), but I guess we'll see them if we see them.