Allow players to be simultaneously berserk and mesmerised
The flavor justifications for this seem a bit weak (players were sometimes
surprised by it) and the gameplay was a bit awkward - on one hand it isn't
very fun to be locked out of using your signature Trog ability or a powerful
emergency potion just because a siren happened to start mesmerising you on
the previous turn, while on the other hand sometimes you wanted to be
mesmerised to protect yourself from berserkitis.

I suspect a large factor in berserk and mesmerisation being mutually
exclusive in the past is that their displays in console interfered with
each other; with this in mind this commit also tweaks the console display
of los "flashes" like the constant red used for berserk to take
mesmerisation into account.