Let coglin melee shadows attack with both weapons
Shadow mimics of coglin players get a second attack, wielding a shadow
copy of their offhand weapon (if they have one).

This probably doesn't play nice with shadow positioning if the player is
wielding a polearm and a non-reaching weapon at the same time, but
hopefully that is a niche enough situation not to worry about it for now.