Add ARTP_TROG_MESMERISE, mesmerise on kills plus doubled regen on mesmerise
(Doesn't yet generate - will be used in Trog gifts.)

This property, displayed as TrogMesm, makes killing a monster in melee cause
the nearest monster to start mesmerising you. The idea is that if you kill
one monster in a group, you are likely to become committed to finishing the
fight, which generally fits well with Trog's style of fighting. Unlike the
obsidian axe, there are various ways to break the mesmerisation chain and
run away: fog, terrain, killing the currently mesmerising monster with
non-melee (evocables or throwing), or even waiting out the (lengthy)
mesmerisation duration.

As a perk, this property grants doubled HP regeneration whenever you are
mesmerised. Note that in addition to helping you fight through a long
mesmerisation chain, this also makes Trog's Hand better.

(Like the obsidian axe, the mesmerisation bypasses clarity and stops you from
unwielding the weapon. The description of the obsidian axe property
also now explicitly mentions that it ignores clarity.)