Make coglin shadow mimics wield weapons in random hands
Compared to players, this *almost* never matters, since monsters can't stab
and their weapons are always unbranded. But technically if a player was
wielding an axe and a non-cleaving weapon, it would be optimal to have the
axe in the main hand so the shadow could always cleave even if the main
target died.