Let ranged bump attack work with sight range 0 (elliptic)
Currently, if the player has a sight range of 0 (ie: during the first turns
of Primordial Nightfall), trying to move into a monster will say "You can't
see that place." and consume your turn without doing anything. For more
consistency with melee attacks, this aims to let you try to shoot whatever
you bump into in this case.

Since ranged attacks use beams to create their ranged_attack (and beams
can't be aimed out of your LoS like this), this short-circuits that part
if your vision radius is 0 and makes the ranged_attack directly.

I am not really happy with this implementation and would like to come up
with something a little nicer (though I'm not sure how plausible this is),
but it seems to produce the desired functionality at present.