Rebalance wanderer good staves skill, remove decent staves skill
For a while now, staves skill has been completely unbalanced for
wanderers, as the wanderer background treated a +0 quarterstaff as if it
was equivalent to a +0 falchion/hand axe/mace/short sword, when it is
considerably better. This commit makes a +0 quarterstaff the good
quality reward for staves skill (instead of a +2 quarterstaff).

I have spent some time considering a suitable replacement for decent
staves skill, including a random magical staff, a magical staff
matching possible magic skills, or simply a -2 quarterstaff. But
magical staves are probably a little too niche, and I consider that it
might feel bad to start with a -2 unbranded weapon, even if in reality
it's okay.

So I've just decided to remove decent staves skill as a possibility.
After all, there is no staves option for the backgrounds that get a
choice of a decent weapon (Mo/Be/CA/CK/Wr/Re), so why should there be
for wanderer?

[Some data: from a sample of 1,000,000 HuWns, good staves skill
occurs about 4.2% of the time, and decent staves skill occurred before
this commit about 4.8% of the time. After this commit, other decent
weapon and magic skills are each picked about 5.1% of the time.]