Let freezing/flaming brands trigger rF-/rC- warnings for players (#3859)
If flaming/freezing deal extra damage to the player because the player
has rF-/rC-, print the usual "The fire burns you terribly!"/
"You feel a terrible chill!" messages. I admit that the former doesn't
go too well with the messages "@the_monster@ burns you" that are
printed by the flaming brand itself, but I figure it's important to
keep this message universally consistent.

It's a little bit hacky, but I want to avoid spamming the message log
with resist messages for freezing/flaming, and we can't let this amuse
Xom because that would be way too easy to abuse. (Both of these happen
if check_your_resists is used, the normal way to get these messages.)

Resolves #3859.