Don't let TSO halos make allies easier to hit
TSO offers some obvious support to a general summoner: halo makes summons
more accurate and TSO's blessings will heal and extend the duration of
summons. It feels slightly awkward in this context that it will *also* make
all of your non-divine summons more likely to *be* hit. Let's lean a little
more into generalist summoner support by making this not be the case.