Greatly boost the effectiveness of heal-from-kills at low piety
While the ability to heal HP/MP from suitable kills was nominally given at
1* for worshippers of Mahkleb, Vehumet, and TSO, the chance of it actually
happening on a suitable kill at that point was less than 1%. It basically
didn't *really* exist until a decent bit more piety was gained.

Since the amount of healing itself was already based on the HD of what was
killed (making it naturally weaker in the earlygame anyway), it doesn't
need to be practically useless at first. At the same time, it meaningfully
scaled at piety values above 6* and there's been some efforts over the
years to make piety-based things actually max at the highest level the
player can see.

This commit rescales the trigger chance to be 30% at 1*, maxing at 80% at
6*. This is slightly below the old theoretical max (85% at 200 piety), but
a combination of piety decay and diminishing returns on piety gain and
actually *spending* piety on things meant that one wasn't likely to ever
stay at 200, and so this is hopefully - at most - a tiny reduction in
practice. (This may be slightly less true for Vehumet, who doesn't have any
way to spend piety, but I still doubt it's significant. If so, these can
be decoupled from each other.)