Don't let the player take actions when their turn is over
If the player entered multiple mouse inputs in a single frame in local
tiles, they would all be processed and the time taken for the player's
action would be set to the time for the last action. The monsters and
environment would then get to react.

Equipping a weapon as a coglin doesn't happen in a single turn but
rather adds a "delay" to a queue that makes you skip your turn until
it finishes equipping or gets interrupted. So, if you start equipping
the same weapon twice in the same turn with this bug, it would skip
your turns until the weapon equipped and then skip your turns until
the weapon equipped again which would crash as you were already
wielding the weapon.

The was another bug were the function to check if there was more input
keycodes (a keycode being e.g. a, D, ctrl-b) was incorrect and would
often return true when ctrl, shift or alt were held even though they
don't give a keycode on their own. This would sometimes cause the game
to freeze when reading in all the available keycodes when one of these
keys were held making it easy to input multiple mouse inputs in one
frame.

The previous two bugs were the cause of issue #3563 and should be fixed
by this change.

I also fixed #3804, #3802, #3801, and #3799 when I was rewriting the
UI code.