Place monster non-adjacent to player at start of Duel (elliptic)
It could be extremely advantageous to melee characters to always start in
melee range of the target they picked (while simultaneously being worse for
ranged characters, which Okawaru also explicitly supports). Now instead the
duel target will be placed in a random spot in the arena, 3-5 tiles away
from the player.

The smallest arena wasn't actually large enough to do this if the player
started at its centre, so we now randomize your arrival location to one
of the corners there instead. I added a small bit of similar randomization
to the second-smallest. The larger arenas remain unchanged.