Allow manifold assaulting with a weapon that warns
Because manifold assault chooses its targets at random, it has to avoid
targets that would warn the player so as not to put them in penance etc.
However, when using a weapon that would warn when attacking any monster
(e.g. the singing sword with an ally in view) this lead to it not
finding any targets it could hit. To fix this, promt when attacking with
a weapon that would always warn when attacking and only exclude targets
where the actual target is the problem (e.g. to close to an ally when
using devastator).

Fixes #3838