Split Black Mark (the spell) away from Black Mark (the mutation)
Black Mark was the signature spell of the now Demonspawn Soul Scholars, but
was usually unimpactful in practice and symmetry with the player demonspawn
mutation of the same name limited its ability to be buffed without
confusing people (and had historically resulted in its most significant
effect being outright removed). An always-on player passive and a signature
short-term status used only in extended have such different balance
considerations that I think it better if they just be two entirely
different effects.

The DS mutation remains named Black Mark (and is unchanged). The Soul
Scholar spell has been renamed to Sign of Ruin and its functionality
changed (though with the same core mechanical thrust). The spell now places
a debuff on a 5x5 area of enemies that will trigger each time those enemies
are hit with a melee attack, applying heavy draining and then one of either
slow, weak, or blind. This is basically an inverted (and more powerful)
version of their old effect, with the antimagic removed (Warmongers in the
same zone already have the anti-magic end of things as a focus).

As a debuff, it can follow the player through an extended battle, but there
are additional lines of counterplay if things get rough enough (ie:
!cancellation works on it). It also works better with the subject of the
immediately following commit.