Adjust Kiku titles and the corresponding Ds enemy
There's not been much adjustment of piety rank titles in the past decade
unless the gods in question were adjusted, which has some silly and rather
strange overlaps in contrast with one another as more and more titles are
racked up between them. Kiku's is one of the strangest; we've got a 1*
"Scholar of Death" as a stepping stone to greater power in contrast to
Sif's 4* "Scholar", a 2* "Merchant of Misery" who has no connection to
Gozag or any selling in Crawl, a 3* "Artisan of Death" who doubles up
on the word death just in case one missed the theme of the god, a 4*
"Dealer of Despair" unrelated to Nemelex, and a 6* "Lord of Darkness"
unrelated to Dith. Most jarring is the occult symbol title at 5* that
even beyond its unfortunate modern associations breaks the naming
schema in nearly every way: neither "X of Y" nor a pairing of concrete
job and suffering.

This adjusts some (if not all) of Kiku's piety titles to focus a little
harder on a particularly academic angle of necromancy (helping further
distinguish from Yred's more ferverent faith angle), as well as deal
with a variety of the prior mentioned matters.

This naturally also comes with renaming (and retiling) the Demonspawn
enemy that explicitly invokes Kiku and mirrors one of the available
titles. These new demonspawn soul scholars are a little less dramatic
in their new names, but add a bit more texture to all the intimidation
and bluster attached to other unholy matters throughout Crawl (see
also Nergalle and Xak'krixis)- it's also no more bland than the current
corrupters. They could honestly do with a pretty significant buff on
their signature spell anyway (which struggles immensely against extended
player defenses), which would coincide well with this rename.