Replace Kiku's Torment power with Sign of Ruin
Kiku's Torment ability was widely considered a very niche aspect of their
kit. While its effect could be situationally powerful, it was unpredictably
dangerous to the player *and* duplicated a scroll effect. There were often
already enough scrolls of torment in a game for the number of times one
really wanted to use torment (unless the player was undead) resulting in it
just not being a button people were very excited to have. So let's
experiment with something else.

The revamped Sign of Ruin was done with this possibility in mind; Soul
Scholars were always Kiku worshippers, after all. By debuffing a group of
enemies with something that scales directly off receiving melee attacks
(and has no friendly-fire), it plays well with having lots of minions in
play, but importantly does not actually *require* them. (An earlier draft
used the old ally-buff version of Black Mark, but if Kiku hadn't deigned to
give you actual ally creation spells you could use with Wretches, it was
very sad).

Kiku has the highest abandonment rate of any god besides Ignis (part of why
their wrath was buffed significantly a short while ago). Maybe a little
carrot to go with that stick?

(The numbers are still pretty fuzzy, but I'll adjust them based on feedback
if the net effect of having this ability feels good.)