Fix delay interrupts sometimes prompting twice
In delays' try_interrupt method, the code prompts the player as to
whether they would like to interrupt the delay with the yesno function.
However, before this is resolved, the yesno function itself can
(eventually) call try_interrupt a second time (and no more times,
because at that point there is a mechanism to block recursive
interrupts). Thus, in order to avoid a double prompt, it's necessary to
record that the player has been prompted before yesno is called for the
first time.