Don't crash on reusing a shadow mimic with a non-ranged weapon
If a player created a shadow mimic while wielding a non-launcher, then
switched to a launcher and triggered a shadow shot, the code would assume
it could reuse the existing shadow mimic and attempt to fire from whatever
it was wielding (which crashed). Now we simply don't reuse a shadow mimic
for shadow shoot unless it's wielding a launcher.