Fix all statuses coloring SH blue in webtiles
After removing Divine Shield's effect on SH, this left SH-boosters with an
empty string (even when there is no defined "sh" member at all...) which
matches all other strings and made all statuses color your SH.

I could reorder checks so that it never looks for "sh" here, but then it
might surprise someone in future that tries to put another in and it
doesn't work.

This is still dumb, though.