Cut an extremely niche and mostly-pointless monster behavior
If holies walked over a good god altar, or branch natives of orc, elf, or
spider (but no other branches!) walked over the entrance/exit to their
branch (only the entrance, in Spider's case) while wandering, they had a
chance to set that location to their patrol point and stick near it. Did
you know that?

Of course you didn't.

(Realistically, this mattered slightly more when monsters could spawn on
top of stairs over time, and this could cause these monsters to cluster up
around the place they came out of. But that behavior is itself long gone,
so this can join it.)