Quick stopgap measure to fix webtiles flashes sometimes being fully opaque
For unclear reasons that I have been completely unable to reproduce,
flashes for some players in webtiles sometimes become completely opaque
instead of transparent. Even worse, since this applies to the constant
screen effect from berserk, it can make the player completely unable to see
the battlefield for the entire duration of the effect.

This behavior is inconsistent, and seems to sometimes go away on reloading,
but given that we can't expect players to know or do this, I've attempted
to disable the only place I can see that allows for variable alpha
transparency in flashes (ie: any value other than hard-coded 100/150).

This will result in the screen effect for blind being a bit overly loud,
but hopefully this is just a short-term measure until someone can figure
out what the real problem is. (I mean, hopefully it even works - it's hard
to test a fix to a problem I cannot reproduce...)