Replace Necrotise with Soul Splinter
Necrotise is a perfectly functional damage spell for early floors, but the
skeleton often feels like an afterthought; there's frequently nothing
around for it to hit and it will usually be gone by the start of any
subsequent battle. By the time this is less true, its damage is already
starting to fall off. (Also: neither it nor Vampiric Draining can do
anything to rN+ monsters, some of which are quite threatening to a
background that is already considered a weaker mage start).

There has been some desire for necromancer to have a more reliable 'ally'
spell before reaching Animate Dead, but one that still clearly falls off
past earlygame. Derived undead have some issues in that regard, with how
they naturally scale with the monsters they are made from, so this is an
attempt at accomplishing this goal with a new fixed ally.

Soul Splinter is a will-resisted enchantment that breaks off a tiny
fragment of a monster's soul into a Soul Wisp: a short-lived ally that will
fixate on the monster it was extracted from (much like Haunt summons do at
higher levels) though can stick around a short while afterward. A wisp can
be created from any spectralisable monster and will spawn adjacent to it,
also inflicting weakness on the monster for more or less as long as a part
of its soul is out of it. Higher HD monsters will produce wisps with
somewhat stronger attacks, but its health is always low.

The beam-based targeting and unique monster placement should make it feel
distinct from existing level 1-2 summons while remaining necromantic. It
notably helps necromancer's bad matchup against ufetubi. The fact that it
can spawn on the opposite side of a monster in a corridor to let the
necromancer team up on someone is also relevant. Let's give it a try!