Add Grave Claw (level 2 Necromancy spell)
This is partially intended to make Vampiric Draining no longer the singular
damage tool a necromancer relies on before Animate Dead comes online, and
also as a relevant replacement for Cigotuvi's Dreadful Rot in the Hedge
Wizard book, where I think it may also play interestingly.

Grave Claw is a smite-targeted direct damage spell that does physical
damage and inflicts a very short-term Bind effect on a monster. It can be
fired while on the other side of your allies (allowing it to play well with
Soul Wisp - and sometimes help it get an extra hit on a monster if the wisp
spawns on the other side of it), can damage rN-immune targets, and even
does more upfront damage than Vampiric Draining (though checks AC). With
one important stipulation: it cannot be recast until you have killed 2-3
living/zombifiable monsters. This makes it a situational tool that can
help equalize otherwise difficult fights, but can never be the main way you
deal with anything.

I think a spell directly relating to harvesting deaths is good flavor
for early necromancers to have, especially now that their level 1 spell no
longer does this. It helps with hostile undead (while not preventing them
from being an overall bad matchup) and I think it presents some interesting
decisions. It does overlap with Vampiric Draining in several ways, but I
think there are enough distinctions that this will improve the gamefeel of
their kit as a whole.

(Hedge Wizards can also benefit from a more impactful spell in their
necromancy slot, and they too have an ally to synergize with it.)