Scale Xom's summons to XL, introduce a lot more options
For a god of chaos that has little restraint in choice or power, Xom is
kind of dull in their choice of servants. Demons tiers 3 through 5 and
some random shadow creatures both binds singularly demons too heavily to
chaos and leaves out a variety of more fun and weirder options than what
one will see plenty of in Orc or the branch they're already within.

Alongside the old demon summons (and a few more demon summons at higher
tiers), this adds a fairly large amount of various living and undead
monsters themed around various themes Xom already focuses upon:
relocation, confusion, mutation, variable summoners, chaotic natures, and
so on and so forth. There's also a sparse amount of additional demons
meant to preserve and extend the old demonic flavour, mostly aimed at
demons not more obviously used by Makhleb or Nemelex. To help make it
more plausible to parse such, there's also a decent chance to make given
choices appear in pairs. Also, shadow creatures calls themselves will
always get one of these manifold chaos options (and are slightly less
common), to both make it a little less plain and because it's a little
funny to see a given demon look lost in a swarm of Vaults humans and
troll / taurs.

This also adjusts the scaling of monsters chosen and placed: it is now
impossible to meet any tier-3s on D:1 without Xom being bored or
abandoned, while one will stop seeing tier-3s past lair branches. There
is also a lot less of a chance to see over 6 or so monsters, outside of
a couple of banding-designated monsters such as laughing skulls or dream
sheep, since the overall pool is a lot more dangerous / strong later on
than it used to be. This should mostly make it so Xom is much less likely
to directly kill on the first few floors, but is always providing some
degree of potential threat later on instead of hitting a tier-3 wall or
solely relying on shadow creatures. While summoned allies are also scaled
to be notably strong later on, getting more than one comes with quite a
short duration by default, so it's just a per-fight uncontrollable perk.

TODO: This is still somewhat rather limited in decent options for lategame
and extended characters, in trying to not overlap too hard with Lugonu,
Makhleb, Jiyva, or using holies. Maybe Xom briefly summoning a distorted
bound soul of a slain unique as an ally could pass as a joke at Yred's
expense? Maybe we just need a few more Depths, Zot, and Pan enemies.