Make simple god and monster msgs apostrophise.
By adding an extra parameter to the functions, so that "'s" isn't
hardcoded everywhere the possessive is needed.

The only instances of a hardcoded "'s" left are ones in string
replacements that will likely break if changed (or if apostrophise()'s
current logic is changed); the vast majority are now properly passed
through apostrophise().