Fix a weird infinite loop with ally retreat orders (robotcentaur)
Due to an oversight, after ally retreat behavior was refactored by 6316ba047ae8947d194a466133b87ee628c4d497
it didn't actually properly end once they'd reached their destination.
Their behavior and patrol points would be set properly to end it, and then
handle_behavior would silently reset their behavior back. I am frankly
astounded that no one else reported this problem before now.

But even worse: because it would set their patrol point to (0, 0) WITHOUT
actually changing their behavior to not seek it out, anything that reached
the end of where it was going would immediately attempt to travel to
(0, 0). And apparently if you were on a level where the ally in question
could get near to the level boundary (eg: Shoals), this would cause an
infinite loop. I am not sure the specific cause of this, since it went away
when the first bug was fixed; I assume it was because handle_behavior kept
getting called to give the monster somewhere valid to move and it kept on
insisting it should move to (0, 0).

Anyway, it seems to work properly now.