Cigotuvi's Dreadful Rot -> Cigotuvi's Putrefaction (level 4 Necro/Air)
Cigotuvi's Dreadful Rot was a spell in an awkward place. It could be used
to open up small gaps against enemies in corridors in the earlygame, but it
was found only on the background which already had Call Imp and Blink, both
of which can do this more flexibly (and actually be a primary kill tool, in
the case of Call Imp). You could sometimes use Mephitic Cloud to confuse
enemies into it, but it was unclear how much help you needed against an
enemy that was already confuse-locked.

So let's try a different and higher-level take on making miasma clouds.

Cigotuvi's Putrefaction is a smite-targeted level 4 Necromancy/Air spell
which must be cast on a living monster that is at least 'heavily wounded'.
Repurposing the old cloud spreader code (last used about 10 years ago, but
never removed) it causes miasma clouds to billow out from this monster over
the next several turns. These clouds start as harmless 'faint miasma' that
exists to give the player a better idea of the area of effect before
turning into proper miasma the next turn. Faint miasma will never turn
harmful so long as the player themselves is standing there (so no matter
what terrain you create the clouds on, they won't hurt you - though it's
still possible to surround yourself!)

The spell has a fairly significant draining cost, but this cost itself is
heavily reduced by spellpower. (Increasing number of clouds and cloud
duration via spellpower felt like it ran the risk of sometimes making the
spell harder to use instead of strictly better.)

Let's see if Corpse Rot the Fifth (Sixth?) feels in a better place.