Make Vehumet's 2nd gift be level 2-3 instead of 1-2 (elliptic)
There's already a fair bit of randomness in when the player can be offered
Veh gifts (compounded by randomness in when one finds Veh altars in the
first place), and getting a level 1 spell for your second gift is almost
universally sad - even if worshipped fairly early. Let's be a little more
generous.