Fiery and icy Swamp end vault fusion
I've never been much of a proponent of leaning on threats not based on a
branch for branch ends when said branch ends are supposed to be the
dramatic finishing point of the branch in question. It dilutes the
branch identity and threats already present in a branch to swerve over
into a different direction with only the loosest of collective theme
justifications, and frequently doesn't even try to connect them back
to the monsters present. If we need more branch end variety, plenty
can be done while still trying to connect to what's present and with
new layouts. Sometimes this guests-work has been retroactively made to
make more sense (like pharoah ants in Spider for the death scarab end),
but I'd like to continue to keep a stern eye on such spaces regardless.

One of the most obvious manifestations of this is in Swamp, which despite
getting notable enemy updates in 0.14 and 0.26, has had a historical
obvious emphasis on looking elsewhere for more threats (due to design
patterns built up from a decade and a half ago prior to either update).
I've pared this back down a little over time for other concerns, and will
continue it here by merging two other troublesome vaults- swamp_fiery and
swamp_icy. These two have had issues for a while- the former's
unrestrained clouds fry most of what's inside before being fought (and
for a long while was at the bottom of threat level due to using, like,
steam dragons and fire crabs for a rune vault), while the latter relies
heavily on slow melee monsters and has extremely little connection to
Swamp itself beyond modified hydras. This merger emphasizes a few
different points:

 * Having both themes provides enough variety to still work with both
   the old iconic monsters and regular Swamp spawn. The half of present
   monsters that aren't hydra simulacra are instead branch spawns that
   still work with simulacra, like swamp dragons and will-o-the-wisps.

 * The clouds and walls have been reworked to make the former more
   organized and the latter thicker, enough to reduce the likelihood
   monsters will wake up and wander into damaging internal flaming clouds
   before actually approaching the vault. It also makes the external
   freezing clouds mildly easier to move past, to focus more on other
   internal threats.

 * In an experimental difference from other swamp layouts, combining
   both layouts is meant to emphasize several large open chambers
   rather than dense singular cave-boxes or a long corridor of spawns.

Let's see how such revisions work out. The old two vault layouts might get
used eventually for more elemental game-long vaults later on, but the
vault's sort of combining the two shapes together, so it's fine for now.