Add a few more Kiku Donald/player ghost lines.
To reflect the flavour change from pain and torment to ruin and despair.

The player ghost line referencing pain and torment is still there, but
the way it's worded, it could apply to the ghost instead of the player,
and who knows how Kiku punishes worshippers who fail them by dying?