Improve Step From Time's functionality
Step From Time is a conceptually cool ability - let the level run for a few
hundred turns without the player and see how much that improves your
situation - but its reliability is deeply troubled in practice. It is very
common for it to make no meaningful improvement in your situation - and
worse, enemies get an immediate action so sometimes the player can even
get blasted by monsters that weren't in sight before they used it!

A fundamental problem with the ability working the way it does is that
monster wandering behavior in Crawl - when there is no player to track nor
noise to pursue - is very, very dumb. A monster can literally spend
thousands of turns in a medium-sized room without figuring out how to exit
it - a problem that gets even worse the more other monsters are in the room
for it to bump into, as each time its movement is blocked, it will start
moving in a different random direction instead, making it near-impossible
for a cluster of monsters to exit a corridor.

I spent a while investigating how this basic wandering behavior might be
improved, but it's unclear that any changes which do not require a
massive-scale refactoring of monster behavior code (which is a good idea in
general, but very daunting) would actually be beneficial. The incidental
property of 'monsters tend to stay inside enclosed major vaults unless the
player is fighting nearby' is arguably good, more intelligent behavior
could be much more computationally expensive, and what monsters do when the
player is extremely far away from them mostly just isn't that important.

Except for Step From Time.

So instead, let's cheat. Rather than merely causing a large amount of time
to pass and hoping the results incidentally make a good button to press,
let's move some things behind the scenes to ensure that they do - while
still staying true to some of the flavor and 'unpredictability' of the
ability - but to a more acceptable degree.

Step From Time now passes only 100 turns in the old way (enough to
introduce some general randomness, and cause timed effects to wear off,
summons expire, poison to kill things, etc.). And then it grabs a random
50-85% of nearby awake monsters and manually moves them some distance away
from the player to simulate them wandering (if wandering actually worked
any good at all). They will always end up somewhere out of the player's LoS
that they could have walked under their own power (so it's definitively
*not* a teleport).

Then we roll an independent power-based chance to put awake creatures
anywhere in the level to sleep. (A property Step From Time effectively had
in the old days, and one that seems thematically appropriate without being
too easy to exploit.)

Finally, we steal just enough energy from monsters which remain in the
player's LoS to ensure that they do not act before the player has seen
them.

In total, using this ability is a lot more likely to improve the player's
situation (but is not guaranteed, still has a large degree of
randomization, and still functions poorly in places the player could
*expect* it to function poorly, like enclosed spaces with no exit.)

I've nudged the piety cost up slightly at the same time, but it should feel
like a much better button to push, regardless.