Indicate whether species/forms have blood.
In forms, move the quality from can_bleed to has_blood. Elsewhere,
rename the function. In species, make a new species flag for it.

Note that having certain holinesses no longer automatically implies
not having blood. Species with no blood are explicitly flagged as having
no blood instead. (The one exception to this is the check for
is_lifeless_undead()'s being false, which is needed to account for
the one odd case of not-currently-alive vampires.)